Moved load to saveLoadController
